add mudule estructure
