+++ index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./assets/css/style.css">
+    <link rel="stylesheet" href="./index.css">
     <title>Awasome Books</title>
 </head>
 
@@ -12,7 +12,7 @@
     <header>
         <div class="navbar-container">
             <div>
-                <h2 class="logo"> <a href="#form-area">Awesome Books</a></h2>
+                <h2 class="logo"> <a href="index.html">Awesome Books</a></h2>
             </div>
             <nav>
                 <ul class="list-nav">
@@ -44,9 +44,9 @@
             <p>Do have any questions or just want to say "Hello"?</p>
             <p>You can reach out to us!</p>
             <ul id="list-contact">
-                <li>Our e-mail: email@mail.com</li>
+                <li>Our e-mail: gracianomanuelhenrique@mail.com</li>
                 <li>Our phone number: 004354375754</li>
-                <li>Our address: Streetname 22, 84503 City, Country</li>
+                <li>Our address: Streetname 22, 84503 City, EUA</li>
             </ul>
         </section>
     </main>
@@ -56,7 +56,7 @@
             <p>Copyright of the Graciano Henrique © 2023</p>
         </div>
     </footer>
-    <script src="./assets/js/main.js">
+    <script type="module" src="./index.js">
 
     </script>
 </body>
