+++ module/bookLibrary.js
@@ -1,0 +1,5 @@
+export class BooksLibrary {
+  constructor() {
+    this.bookArray = (JSON.parse(localStorage.getItem('booksDb')) !== null) ? (JSON.parse(localStorage.getItem('booksDb'))) : ([]);
+  }
+}
