+++ module/variable.js
@@ -1,0 +1,10 @@
+const btnAdd = document.querySelector('#add');
+const authorInput = document.querySelector('#authorBook');
+const titleInput = document.querySelector('#titleBook');
+const listLink = document.querySelector('#list');
+const contactLink = document.querySelector('#contact');
+const addLink = document.querySelector('#add-new');
+const h1 = document.querySelector('#h1');
+const date = document.getElementById('date');
+
+export { btnAdd, authorInput, titleInput, listLink, contactLink, addLink, h1, date }
